Date_Entry_16Oct19
I'm close. In the handleScreenTouch function I added "datePressed = 1" this is set when I press date on the screen. I then change the if statement in handleKeyboardPress to test that datePressed is true. If it is I call updateDateItem. updateDateItem is how you wrote it plus I added  "datePressed = 0"

It now waits until I press date on the screen and shows me what I entered after I hit Enter. I added the Serial.print(currentDate[x]) as you mentioned to confirm it's working. I also added tft.print(currentDate[x])  to show the date on the screen but that just shows the last character of what I entered. I believe both print statements are printing each character as it's read from the buffer just displays different.

So I think all I need is to convert those characters into a complete string to show on the screen